+++ src_arduino_backup/beat_detection.h
@@ -1,12 +1,12 @@
 #ifndef BEAT_DETECTION_H
 #define BEAT_DETECTION_H
 
-#include <stdint.h>
+#include "driver/i2s.h"
 
 #define SAMPLING_FREQUENCY_HZ 48000
 #define FFT_BUFFER_LENGTH 1024
 
-extern uint32_t lastBeatTime_ms;
+extern unsigned long lastBeatTime_ms;
 extern bool isBeatDetected;
 
 void ComputeFFT(int32_t rawMicSamples[FFT_BUFFER_LENGTH]);
